Added media message

+++ DS-Chat-CS1/DS-Chat-CS1/Model/ImageMessage.cs
@@ -17,8 +17,8 @@ namespace DS_Chat_CS1.Model
                 RaisePropertyChanged("ImageUrl");
             }
         }
-        private int _loading = 0;
-        public int Loading
+        private double _loading = 0;
+        public double Loading
         {
             get { return _loading; }
             set
